requirements.txt
Flask>=2.2.5
Flask-Login>=0.6.2
mysql-connector-python>=8.0.33
Werkzeug>=2.2.5
